[BUGFIX] Report module should not alert when all db tables are utf8
In case a database default character set is latin1 but all tables
created in the database are utf8, the reports module warns that
the db is configured wrong, which might be the case, but on certain
hosters this is not an excuse. Instead, all TYPO3 tables
are created with a utf8 collation (equiv. to utf8 character set) and
stored correctly, so there is no need to throw a big red box.

The patch extends the utf-8 check to see if all database tables
are utf8 and everything is fine.

Resolves: #78413
Releases: master
Change-Id: I78cc65e4fec58e40f41f36ec6dbf6d46e867246f
Reviewed-on: https://review.typo3.org/50351
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>