[!!!][FEATURE] Replace SwiftMailer with symfony/mailer and symfony/mime
SwiftMailer is not in active development anymore
as the author created the successor symfony/mailer
and symfony/mime packages to create and send
emails.

This is a breaking change, as PHP mail() is not
supported anymore. This is now automatically
switched to "sendmail".

In addition \Symfony\Mime\Email has a different
signature than \Swift_Message, which will
cause some trouble, however we've managed
to overcome most of that functionality to
stay backwards-compatible.

Also, all extensions extending from SwiftMailer
will fail as the package is be removed
with this patch.

Spooling has been reimplemented as direct transport
methods (DelayedTransportInterface) instead of
Symfony Messaging for the time being.

Used composer commands:
 * composer require symfony/mailer symfony/mime
 * composer remove swiftmailer/swiftmailer

Resolves: #88643
Releases: master
Change-Id: Ic02db633392f1d0d7b7061e7b322435f892d2c04
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61152
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>