[TASK] Set ModuleTemplate details from request
This change adds a new ModuleTemplateFactory, which
is built with the Request in the object, allowing to use
the current request of the current action.

At the same time, all TYPO3 Core controllers
use proper DependencyInjection now, making it easier
to spot the dependencies and uses PSR-7 Request
where possible.

Resolves: #93885
Releases: master
Change-Id: Ia6594a1e68c5c27375f34e49bffeda4dcbc980d7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68729
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>