[FEATURE] Allow icon identifiers in the report module registration
To further streamline the usage of the IconRegistry, the reports
registration array now allows to define an icon identifier. Absolute
paths and paths with "EXT:" prefix are still possible.

Further improvements:
The IconViewHelper of the reports extension has been dropped. It had
the same functionality as a normal ImageViewHelper with the addition of
a fallback icon. This logic has been moved to the controller action.

Some minor code refactoring and template cleanup while at it.

Resolves: #94588
Resolves: #94590
Releases: master
Change-Id: I901b8a0c56f26459c609a36a59c67b341c6b4168
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69965
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>