[TASK] Update display of add/edit scheduler task form
Previously, the form fields in the add/edit form of a task were
displayed in their entirety, filling the screen. Additional settings
were placed at the end of the form, so that changes to this area
were usually not visible when the task type was changed.

With this patch, the width of the input mask is now reduced in order
to increase the clarity within the form and to create a uniform
appearance with forms of the form engine. The markup of the
checkboxes has been slightly adjusted, the position of the button for
selecting a start page ('LinkValidator' task) has been corrected and
individual spacing between the form fields has also been adjusted to
match those of the form engine.

In order to better guide the user in the setting steps, the fields
are now grouped into form areas according to their purpose.
Likewise, the fields for advanced settings are moved to a higher
position so that changes in this area are more visible when changing
the task type.

The task type is now also displayed in the edit form.

In general, this patch aligns the add/edit task rendering of the
scheduler module more with FormEngine rendering. Mid-term, this
module should use FormEngine for rendering anyways, so this is
a kind of a pre-patch.

There is a minor glitch when single tasks don't provide additional
fields, now having a lonely heading. We may change this with a
dedicated patch since it's a bit tricky with the current javascript
hide/unhide implemenation of additional fields.

Resolves: #100042
Releases: main
Change-Id: Ia8b16c07167d5b368a31ba66be5db3d96481a809
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77977
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>